perf: use a Character struct, add WidthHint to Text
Add WidthHint to the Text struct to provide a hint to the renderer how
wide the character is. In many cases (via the Print methods) we are
already getting width for free in our call to Characters. Return a
Character struct which contains the width information, pass this on to
the renderer.

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>