vaxis: add windows support
Add Windows support. Modify Signals and handling of opening /dev/tty to
better handle windows APIs. We instead open any of Stdout, Stderr, Stdin
- as long as it isn't redirected. We hope one of them isn't, otherwise
we fail. Because of this, we can no longer close the tty when we are
done - so we only attempt to open the tty ONCE. Our parser continues
to run even in a suspended state (though the foreground process will
read first so we are ok).

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>