widgets(term): give up on mode 2026
To make 2026 work we'd need to double buffer the screen since the host
application can clear it without us knowing. We'd have to keep the last
rendered terminal screen and draw onto a new one, then swap the screens
when 2026 is turned off. We can drop all this and deal with some
flickering...maybe someday we can do this but it's not necessary for now

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>