term: support OSC 11 queries
Add support for the OSC 11 control sequence queries to the embedded
terminal. Programs can use this to query the current background color.

This is implemented by introducing a new Vaxis method which queries the
host terminal. The built-in terminal uses this method to obtain the
background color on the fly and reports it back to the program.

This feature was motivated by aerc which uses the terminal to run editor
when composing messages. The editor, such as neovim, will now be able to
query the background color and adapt its colorscheme accordingly.

Implements: https://todo.sr.ht/~rockorager/vaxis/15
Signed-off-by: Karel Balej <balejk@matfyz.cz>