sixel: Greatly speed up resizing by quantizing with an octree algorithm
Instead of creating a palette from a median quantizer, then brute-forcing
the palette by drawing the image again against the palette.

This brings a 10x speed improvement for drawing SIXEL images, at the cost
of a slightly image quality decrease.

As the sole author of octreequant, I also submit the file under the
LICENSE of vaxis.