vaxis: add blocking PostEvent which doesn't drop events
Add PostEventBlocking, which works the same as PostEvent but does not
drop events. This function is only safe to use when called from a
different goroutine than the main thread, otherwise a deadlock may
occur.

We use this internally to post events from goroutines we create so that
we are sure no events are dropped.

Reported-by: Vitaly Ovchinnikov <v@ovch.ru>
Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>