term: don't respond with terminal colors to all OSC 11 queries
OSC11 allows to both query and set background color. The programs
that try to set the color, for example, 'fish' shell, don't expect
the terminal to answer with the current background color. This results
in 'fish' putting the response onto the command line.

So only respond if OSC11 payload is '?', otherwise simply ignore the
command.

In future this can be extended to passthrough OSC11 to the underlying
terminal to change its background color.