sixel: set P2 = 1 for transparency
Foot had a change back in the 1.19 cycle to use the sixel palette for
index 0 rather than the terminal palette. This caused a regression in
sixel rendering where images with transparency would have black boxes.
go-sixel hardcodes the sixel introducer, so we instead just modify bit
at index 4: this is what is currently set to 0x30 ('0') but we want it
to be set to 0x31 ('1') to enable transparency.

Tested on foot and contour