fix(vxfw): ensure widgets satisfy minimum draw context
text, textfield, and textinput all customize their surface size and have
the potential of returning a surface smaller than the minimum supplied
in the draw context, which can mess with layouts.

This commit also includes a test that renders each widget with both
short and long content and checks the returned surface.