fix(vxfw/list): allow list elements to be size 0
After a flex layout computes widget sizing, it re-renders each widget
with the min and max set to the same value. In the case of a list, this
tight constraint was being passed to the children, and then each child
was rendering itself with min==max, resulting in *each* child taking up
the full flex area.