sixel: Fix reported race condition when drawing a resized image
go -race reports a race condition when drawing a resized sixel image,
because the check for s.encoding is done after we check s.buf.Len(),
which is being modified by the resizing routine.

This fixes the issue by checking for an encoding image first, then
for an empty one.